test: update Similar Item tests for new package
