--- src/ramstk/controllers/similar_item/__init__.py
@@ -1,9 +1,0 @@
-# pylint: disable=unused-import
-# -*- coding: utf-8 -*-
-#
-#       ramstk.models.similar_item.__init__.py is part of The RAMSTK
-#       Project
-#
-# All rights reserved.
-# Copyright 2007 - 2020 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
-"""The RAMSTK similar item controllers package."""
